extract colab_utils to utils so runtime can depend on it without a cycle
