+++ mesop/colab/colab_run.py
@@ -2,12 +2,12 @@ import threading
 
 from absl import flags
 
-from mesop.colab import colab_utils
 from mesop.runtime import enable_debug_mode
 from mesop.server.constants import EDITOR_PACKAGE_PATH, PROD_PACKAGE_PATH
 from mesop.server.logging import log_startup
 from mesop.server.server import configure_flask_app
 from mesop.server.static_file_serving import configure_static_file_serving
+from mesop.utils import colab_utils
 from mesop.utils.host_util import get_default_host
