+++ mesop/colab/colab_run_test.py
@@ -3,9 +3,9 @@ from typing import Any, Callable
 
 import pytest
 
-from mesop.colab import colab_utils
 from mesop.colab.colab_run import colab_run
 from mesop.runtime import runtime
+from mesop.utils import colab_utils
 
 
 class FakeThread:
